# docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API do sistema GuardPoint (controle de rondas/postos).",
        "title": "GuardPoint API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/alertas": {
            "get": {
                "description": "Para supervisores, os alertas sao automaticamente filtrados pelos postos aos quais estao vinculados (via posto_supervisores). Admins veem todos os alertas. Os filtros de query sao adicionais — posto_id pode ser usado para refinar a busca.",
                "tags": [
                    "alertas"
                ],
                "summary": "Lista alertas com filtros e paginacao (admin/supervisor)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Status do alerta (aberto, reconhecido, encerrado)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Tipo do alerta (atraso, no_show, senha_emergencia, senha_customizada, sabotagem)",
                        "name": "tipo",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "ID do turno (uuid)",
                        "name": "turno_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "ID do posto (uuid)",
                        "name": "posto_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limite de itens (max 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset da paginacao",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.AlertaListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/alertas/estatisticas": {
            "get": {
                "tags": [
                    "alertas"
                ],
                "summary": "Estatisticas agregadas de alertas (admin/supervisor)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.AlertStatistics"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/alertas/{id}/encerrar": {
            "put": {
                "tags": [
                    "alertas"
                ],
                "summary": "Encerra um alerta (admin/supervisor)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do alerta (uuid)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.StatusResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/alertas/{id}/reconhecer": {
            "put": {
                "tags": [
                    "alertas"
                ],
                "summary": "Reconhece um alerta aberto (admin/supervisor)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do alerta (uuid)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.StatusResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/biometric/login": {
            "post": {
                "security": [],
                "tags": [
                    "auth"
                ],
                "summary": "Login via dispositivo biometrico registrado",
                "parameters": [
                    {
                        "description": "Credenciais do dispositivo",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.BiometricLoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/biometric/register": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Registra um dispositivo biometrico para o usuario autenticado",
                "parameters": [
                    {
                        "description": "Dados do dispositivo",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.BiometricRegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.BiometricRegisterResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "security": [],
                "description": "Admin e supervisor autenticam com \"email\" + \"senha\". Vigia autentica com \"codigo_empresa\" + \"nome\" + \"senha\" (nao usa email). Enviar exatamente um dos dois pares de credenciais.",
                "tags": [
                    "auth"
                ],
                "summary": "Login por email+senha (admin/supervisor) ou codigo da empresa+nome+senha (vigia)",
                "parameters": [
                    {
                        "description": "Credenciais (email+senha OU codigo_empresa+nome+senha)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Encerra a sessao do dispositivo atual",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.MessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/refresh": {
            "post": {
                "security": [],
                "tags": [
                    "auth"
                ],
                "summary": "Renova o access token a partir do refresh token",
                "parameters": [
                    {
                        "description": "Refresh token",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.RefreshRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "\"email\" e obrigatorio para role \"admin\"/\"supervisor\" e opcional para \"vigia\". \"nome\" precisa ser unico dentro da empresa (pode repetir entre empresas diferentes); \"email\", quando enviado, e unico globalmente.",
                "tags": [
                    "auth"
                ],
                "summary": "Cria um novo usuario (somente admin)",
                "parameters": [
                    {
                        "description": "Dados do usuario",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/checkins/lote": {
            "post": {
                "tags": [
                    "turnos"
                ],
                "summary": "Processa um lote de check-ins offline (ate 500 por requisicao)",
                "parameters": [
                    {
                        "description": "Lote de check-ins",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.CheckinRequest"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.LoteCheckinResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/config/escalonamento": {
            "get": {
                "tags": [
                    "escalonamento"
                ],
                "summary": "Lista as configuracoes de escalonamento da empresa (somente admin)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.ConfigEscalonamento"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "escalonamento"
                ],
                "summary": "Cria uma configuracao de escalonamento (somente admin)",
                "parameters": [
                    {
                        "description": "Dados da configuracao",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.CreateConfigEscalonamentoRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.ConfigEscalonamento"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/dashboard/summary": {
            "get": {
                "tags": [
                    "dashboard"
                ],
                "summary": "Resumo do dashboard da empresa",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.DashboardSummary"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_guardpoint_guardpoint-server_internal_model.AlertaListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.Alerta"
                    }
                },
                "total": {
                    "type": "integer",
                    "example": 42
                }
            }
        },
...
        "github_com_guardpoint_guardpoint-server_internal_model.BiometricLoginRequest": {
            "type": "object",
            "required": [
                "device_id",
                "device_secret",
                "empresa_id"
            ],
            "properties": {
                "device_id": {
                    "type": "string"
                },
                "device_secret": {
                    "type": "string"
                },
                "empresa_id": {
                    "type": "string"
                }
            }
        },
...
        "github_com_guardpoint_guardpoint-server_internal_model.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "mensagem descritiva do erro"
                }
            }
        },
...
        "github_com_guardpoint_guardpoint-server_internal_model.StatusResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "security": [
        {
            "BearerAuth": []
        }
    ]
}